% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamodules-package.R
\docType{package}
\name{datamodules-package}
\alias{datamodules}
\alias{datamodules-package}
\title{datamodules: Infrastructure for Data Modules and Miscellaneous Data Getters}
\description{
Setup a central location for storing data. Place for miscellaneous 
    data-related functions that for one reason or another don't warrant a full
    package of their own.
}
\author{
\strong{Maintainer}: Andreas Beger \email{adbeger@gmail.com} (0000-0003-1883-3169)

}
\keyword{internal}
