% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikipedia-terrorism.R
\name{wikipedia_terrorism_scrape}
\alias{wikipedia_terrorism_scrape}
\title{Scrape Wikipedia terrorism data}
\usage{
wikipedia_terrorism_scrape(from = 2019, to = NULL)
}
\arguments{
\item{from}{year or "YYYY-MM" string}

\item{to}{year or "YYYY-MM" string, if NULL will default to latest}
}
\description{
Scrape events from the Wikipedia terrorism lists at
\url{https://en.wikipedia.org/wiki/Template:Lists_of_Terrorist_Incidents}
}
